Fix issues with savestates
For some reason PCSX2 doesn't keep track of file handles when using save states, which matters if the game is reading from the host filesystem, which ptr2plus has modified ptr2 to do. This means if the game is streaming from a file when you save a state (like when ptr2 streams background instrumental WP2 files, or cutscene data, it'll typically break when that state is loaded because the file handle either no longer exists, or the file offset pointer is desynced. This commit adds logic to keep track of file handles and save that data in the save state, then when the state is loaded, it reopens those files and sets their file offset pointers to where they were at before.
This appears to fix ptr2, in testing when loading a state there are a few frames of silence, which isn't a big deal, and might be how normal pcsx2 with isos works anyway (I forgot. I haven't tested in edge cases like saving a state while loading a stage yet, might break.